separated manual editing tests from pure remote tests due to how we set them up
